+++ ui/src/app/core/settings.interfaces.ts
@@ -64,6 +64,7 @@ export interface AppSettingsInterface {
   host?: string
   proxyHost?: string
   sessionTimeout: number
+  sessionTimeoutInactivityBased: boolean
   theme: string
   lightingMode: 'auto' | 'light' | 'dark'
   menuMode: 'default' | 'freeze'