+++ ui/src/app/core/settings.service.ts
@@ -26,6 +26,7 @@ export class SettingsService {
   public proxyHost: string
   public formAuth = true
   public sessionTimeout = 28800
+  public sessionTimeoutInactivityBased = false
   public uiVersion: string
   public theme: string
   public lightingMode: 'auto' | 'light' | 'dark'
@@ -64,6 +65,7 @@ export class SettingsService {
     const data = await firstValueFrom(this.$api.get('/auth/settings')) as AppSettingsInterface
     this.formAuth = data.formAuth
     this.sessionTimeout = data.sessionTimeout
+    this.sessionTimeoutInactivityBased = data.sessionTimeoutInactivityBased
     this.env = data.env
     this.host = data.host
     this.proxyHost = data.proxyHost